abstract factory refactoring
a control takes an abstract factory rather than a view or an input and
an output.

this is a big refactoring here! the problem i had in constructing a
control from an input and output objects is that controls compose other
controls as well, and so a control's ctor quickly explodes in parameters
count.

to remedy that, here controls take one factory parameter instead of the
view parameters they took before. internally they use the factory to get
their views.

an abstract factory is itself an interface of dynamic polymorphism:

 struct abstract_factory {
   virtual ~abstract_factory(){}

a concrete factory implements factory functions:

 struct abstract_factory {
   virtual ~abstract_factory(){}
   std::unique_ptr<snake_input> make_snake_input() = 0;
   ...

 struct sdl_factory : abstract_factory {
   std::unique_ptr<snake_input> make_snake_input() {
     return std::make_unique<sdl_snake_input>();
   }
   ...

so a factory knows to make many different objects, but they're all of
one theme. in the example above, they're all of the sdl theme.