ctors take abstract_factory by pointer
not because they plan to observe the pointee, but because i want to make
sure i don't pass in a temporary. the sdl_factory object must live at
least as long as the game, because it contains, among other things, the
sdl_system object, which initializes and quits the sdl subsystem.