reduce logic in snake::handle_event()
the responsibility of handle_event() is recording relevant incoming
events. that's the meaning of handle_event(). other than thoroughly
irrelevant events there should be no filtering or computation taking
place in handle_event() -- only recording the event, so we know we're
aware of it. the filtering and computation happens in update().

by "other than thoroughly irrelevant events" i mean that if there's a
chance an event may be relevant -- any chance at all -- then we
shouldn't consider filtering it. for example, if the snake reacts to
some arrow key presses some of the times, then handle_event() should
record all arrow key press events. the update() function will later see
an arrow key was pressed and decide whether it is valid or is invalid
and should be ignored. on the other hand, if the snake never reacts to
mouse moves, there is no point in recording mouse moves (or any of the
other nearly infinite kind of events the system can throw at the snake).
handle_event() does no more than this kind of rough filtering.