rm direction::nil
i think that null-like values are not a good thing, because they imply
an "i don't know" sort of value. it's better to be clear when i can, and
eliminate the possibility of an "i don't know" answer to a query about a
value. i don't know if this is alway possibly, but it's very tempting to
try, so let's see if we can live without the direction::nil value.
therefore, if we construct a direction value, we have a direction. it
can be up, right, down, or left -- nothing else -- and so when we have a
direction value the cases to handle are clear and concise (how do you
handle a "no-direction" sort of direction?).

this implies that a living snake must always be moving. so, i removed
the test for the snake initially not moving.

the initial position of a snake is at origin, but as soon as update()
and draw() are called the snake moves. so, when the snake is drawn on
screen, it is at its next position and not at origin.

i can test for the snake initially being at origin, but this would be
done without calling update() and draw(), and i think this test would be
fragile because if the snake_control object doesn't explicitly
initialize the position the position might still be initialized if it
has a default constructor that sets to origin. the initialization might
be undefined, depending on the position initialiation mechanism... so in
other words, this should be a test for the position class, and so i
removed the origin position test from snake_control.

the tests prior to this change relied on (and tested for) the snake
initially being still. for example, tests would move the snake left,
possibly do other stuff, and then check the output. that initial
move-left operation however now is ignored, because the snake initially
moves right. so to remedy that i added another constructor taking an
initial move for the snake to allow specifying a direction other than
right. the simplest approach is to add this extra parameter as the last
constructor parameter, but that meant i had to switch the order of
declaration of the member data so the compiler is happy with the
initialization order being in sync between the header file and the
implementation file.