width int conversion ctor is explicit
for good type safety. the point of introducing the width and height
types is that there might be doubt what things like

 rectangle r{10, 10};

mean. which among the parameters is the width and which is the height?
an int is of vague semantic meaning. a width has more meaning--but if an
int silently converts to a width we are back to the original issue where
the line above compiles and we are not sure what it means.

well actually it was my intent to get here, but i left the constructor
implicit so things compile and my task (introducing a width type) is
over quickly. but now that i have a width type, and the rectangle type
makes use of it, it's time to improve on things.