struct sdl responsible for sdl init and quit
hold the sdl_system resource manager (itself responsible for sdl init
and cleanup), and for handling my (boost signal) quit event.

i've long felt the sdl_factory class is ... off ... in some way. like it
does too much. i couldn't explain what it does in one phrase. originally
i intended it to be a concrete implementation of the output abstract
factory -- i.e., responsible for providing a collection of sdl output
instances. i thought a factory needs no state, since it only provides
factory functions, which merely create instances of objects. but then,
not quite sure of myself, i placed in it a member instance of an sdl
window, and an sdl renderer, and the sdl_system resource handler...

now i think it's okay for a factory to have state, if it needs data in
order to create the instances it's responsible of creating. but i don't
think sdl_factory needs sdl_subsystem to create it's output instances.
it's very convenient to create the factory in one line and get the sdl
subsystem taken care of at the same time -- but it adds architectural
complexity, because the class no longer has one raison d'etre.

so, here i extract the sdl class, which is responsible for initializing
and quitting sdl. internally it uses a resource manager for init and
cleanup, and what it does by itself is listening for the quit event and
translating it to an SDL quit event, SDL_QUIT. because the only way to
quit the program is in main(), that's who will have to handle the
SDL_QUIT event. in fact that's the only reason for translating the quit
event to SDL_QUIT: main() has a loop with one and only one break point:
receiving an SDL_QUIT event.