extract struct sdl_app
to take some of the busywork out of main(). sdl_app is responsibile of
tying the fundamental app mechaniques (event dispatch, sdl business)
with the game control. an sdl_app instance means having an sdl game
control. to clarify, a game_control instance means the logic of the
game, but that doesn't come with the event dispatch and sdl business --
this in fact is a prerequirement for a game_control instance. in
contrast, an sdl_app is in essence the game_control instance along with
all the necessary event dispatch and sdl business packaged, so we have a
game_control ready to go.

quit a mouthful.. should i break this down further? look for a way to
simplify this?