game loop lag catch-up
this from gameprogrammingpatterns.com/game-loop.html, a web page
discussing a technique for a main-loop for a game.

a naive game loop, which is roughly what i up to this point, would be

 while(true) {
   process_input();
   update();
   draw();
 }

here, when i test my game i will experience a certain performance,
which of course depends on my code, but also on the the machine running
the game. if the machine running the game is fast, things are okay. if
the machine is slow, the game's world will be slow, and the user will
experience a slow-running game.

there is not much i can do (that i can think of) for the user to
experience a fast game when they have a slow machine, short of writing a
better performing code. however, what i might do is *simulate* a fast
game that merely *looks* slow to the user. in other words, lead the
user's intuition to the idea that the game is fast alright, it just
looks slow.

think, for example, how a drunk person feels. a drunk person feels slow,
sluggish, his (her) eyelids heavy, falling asleep, and everything around
looks like going on slow-motion. the world around is going at normal
speed, but it just looks and feels slow. as the eyelids slowly shut and
then open again, we might miss having seen something--say, a little bird
hopping 5 meters to the left--but having seen the bird at point a and
then at point b 5 meters to the left our brain will have to do the
mental processing of filling the gaps because our eyes have failed to
deliver that.

the idea here is similar. we keep track of the player's clock each frame
we draw, computing the time it took us to draw the frame, and if we feel
we are not going to hit a target frame rate we strive at we are going to
"cheat" a little by pressing fast-forward on the game to reach the point
where it should be. this "fast-forward" is what simulates the game's
world going at "normal speed" even if the user's machine is slow. it's
as if the user's machine is the one who needs to play catch-up with the
game's world, and not the other way around.

how do we fast-forward the game's world? well, we assume that the
update() operation is fast, and that it's the draw() and process_input()
operations that are time-consuming. then, we repeatedly call update(),
and again, advancing the game state but not showing it, until we feel we
are in sync with the player's clock. then, we draw again, and repeat the
process.

roughly,

                             process_input() <--+
                                    |           |
                                    V           |
                           +-----------------+  |
                       +-> | are we lagging? |  |
                       |   +-----------------+  |
                       |         /   \          |
                       |       yes    no        |
                       |        |     |         |
                       |        V     |         |
                       +--- update()  |         |
                                      |         |
                                      V         |
                                    draw() -----+

you will notice that we don't necessarily update. specifically, if we
are not lagging, we don't update. indeed, at the start we are not
lagging. but as we progress, we accumulate lag. then, we update, and if
all is well we should be able to draw. if we still lag, we will keep
updating, skipping time-consuming draw operations, thus fast-forwarding
the game, until we're good and in sync with the player's clock, not
lagging anymore behind it. then, if all is well, we will hopefully enter
a loop of one update, one draw, one input processing. but of course, if
at any point we lag again -- no problem, just keep updating to
fast-forward the game until we can afford drawing again.

this all means that i must make sure the update() is fast. as fast as it
can be, because we use the update() operation for fast-forwarding the
game. if the update operation is slower than the time slot we budget
ourselves then we will never be able to catch up in our fast-forward,
and the game will crawl to its demise.

for example, if we aim at 60 frames per second, that means 1s/60f, or
0.016666667s per frame. that's the time alloted for a frame loop. in
this time of about 16667 milliseconds we must complete an input
processing, an update, and a draw, to deliver a 60FPS game. If the
update() call takes longer than 16667ms then there is no way the game
will catch up with the lag; the lag will just grow and grow
indefinitely, until the game will virtually halt. not a nice experience
for gameplay.

so, my goal is to keep this in mind and keep my update() operation
performant. of course, i must also keep my input processing and drawing
fast, but really these two should be very thin, with next to no code at
all from my side but just from the graphics library.