fix swallowed moves
rather than a single move_request and move_to state data, maintain a
move_requests list and a move_to data. then, every player request for a
move (via the arrow keys) adds a direction element to the back of the
list. the update() routine then takes the first element of the list and
follows through with it. each call to update() will chop one element off
the list, constantly striving for an empty requests list. when the list
is empty, update() just applies move_to, which as before records the
last move direction.

this way, each frame will move the snake in the same direction as in the
frame before that--unless there is one or more move requests pending, in
which case it will process the oldest request. requests are handled one
by one, one request per frame, until there are none, at which point we
default to processing the last request again.